requirements.txt
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
black==23.9.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
cryptography==44.0.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
filelock==3.17.0
flake8==6.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.8
idna==3.10
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.5
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.5.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.3.3
pyasn1==0.4.8
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.1.0
Pygments==2.19.1
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typer==0.15.2
types-requests==2.31.0.1
types-urllib3==1.26.25.14
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.4
websockets==15.0
